perf(capacitor): rename appId to `saloni`
android was synced
