+++ android/app/src/main/res/values/strings.xml
@@ -2,6 +2,6 @@
 <resources>
     <string name="app_name">saloni</string>
     <string name="title_activity_main">saloni</string>
-    <string name="package_name">io.ionic.starter</string>
-    <string name="custom_url_scheme">io.ionic.starter</string>
+    <string name="package_name">io.ionic.saloni</string>
+    <string name="custom_url_scheme">io.ionic.saloni</string>
 </resources>
